+++ packages/vue-cli-plugin-mpx/commands/serve/index.js
@@ -1,7 +1,14 @@
-const { getCurrentTarget, SUPPORT_MODE } = require('@mpxjs/cli-shared-utils')
+const { getCurrentTarget, SUPPORT_MODE, merge } = require('@mpxjs/cli-shared-utils')
+const fs = require('fs-extra')
 const { serveWeb } = require('./web')
 const { serveMp } = require('./mp')
+const { resolveServeWebpackConfig } = require('../../config/base')
 
+const defaults = {
+  clean: true
+}
+
+/** @type {import('@vue/cli-service').ServicePlugin} */
 module.exports.registerServeCommand = function (api, options) {
   api.registerCommand(
     'serve',
@@ -10,14 +17,21 @@ module.exports.registerServeCommand = function (api, options) {
       usage: 'mpx-cli-service serve',
       options: {
         '--targets': `compile for target platform, support ${SUPPORT_MODE}`,
-        '--report': 'generate report.html to help analyze bundle content',
         '--mode': 'specify env mode (default: development)',
+        '--no-clean':
+          'do not remove the dist directory contents before building the project',
         '--env': 'custom define __mpx_env__'
       }
     },
     function build (args, rawArgv) {
+      merge(args, defaults)
+      if (args.clean) {
+        fs.removeSync(options.outputDir)
+      }
       const target = getCurrentTarget()
-      // 利用子进程转发构建
+      api.chainWebpack((config) => {
+        resolveServeWebpackConfig(api, options, config, target, args)
+      })
       return target.mode === 'web'
         ? serveWeb(api, options, args)
         : serveMp(api, options, args)
