+++ packages/vue-cli-plugin-mpx/commands/serve/mp.js
@@ -1,36 +1,14 @@
 const webpack = require('webpack')
-const { parseTarget, getCurrentTarget } = require('@mpxjs/cli-shared-utils')
-const { symlinkTargetConfig } = require('../../utils/symlinkTargetConfig')
-const {
-  resolveWebpackConfigByTarget,
-  handleWebpackDone,
-  modifyMpxPluginConfig
-} = require('../../utils/webpack')
-
-const resolveMpServeWebpackConfig = (api, options, args) => {
-  const target = parseTarget(args.target, options)
-  // 小程序业务代码构建配置
-  api.chainWebpack((config) => {
-    if (args.env) {
-      modifyMpxPluginConfig(api, config, {
-        env: args.env
-      })
-    }
-  })
-  return resolveWebpackConfigByTarget(api, options, target)
-}
-
-module.exports.resolveMpServeWebpackConfig = resolveMpServeWebpackConfig
+const { getCurrentTarget } = require('@mpxjs/cli-shared-utils')
+const { symlinkTargetConfig } = require('../../utils/symlink')
+const { handleWebpackDone } = require('../../utils/webpack')
+const { resolveServeWebpackConfigByTarget } = require('../../config/index')
 
 /** @type {import('@vue/cli-service').ServicePlugin} */
-module.exports.serveMp = function serveMp (
-  api,
-  options,
-  args
-) {
+module.exports.serveMp = function serveMp (api, options, args) {
   const target = getCurrentTarget()
-  // 小程序业务代码构建配置
-  const webpackConfigs = resolveMpServeWebpackConfig(api, options, args)
+  // 小程序构建配置
+  const webpackConfigs = resolveServeWebpackConfigByTarget(api, options, target, args)
   return new Promise((resolve, reject) => {
     webpack(webpackConfigs).watch({}, (err, stats) => {
       handleWebpackDone(err, stats, target, api)
