--- packages/vue-cli-plugin-mpx/utils/name.js
@@ -1,5 +1,0 @@
-module.exports.getWebpackName = function (api, target, pluginConfig) {
-  return [target.mode, pluginConfig.env, api.service.mode]
-    .filter((v) => v !== undefined)
-    .join('-')
-}
